pyproject.toml
[tool.poetry]
name = "CHANGE_ME"
version = "0.0.0"
description = ""
authors = ["Anthony Almarza <anthony.almarza@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "CHANGE_ME", from = "src" },
]
# https://pypi.org/classifiers/
classifiers = [
    "Private :: Not For Public Release"
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0"
djangorestframework = "^3.11"
django-filter = "^2.2.0"
drf-yasg = "^1.20.0"
drft = "^0.0.0-alpha.1"
acru-l-toolkit = "^0.1.0"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
django-storages = {extras = ["boto3"], version = "^1.11.1"}
django-compressor = "^2.4"
django-environ = "^0.4.5"


[tool.poetry.dev-dependencies]
pytest = "^6"
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.4.0"
pytest-responses = "^0.4.0"
pytest-picked = "^0.4.4"
pytest-instafail = "^0.4.2"
freezegun = "^1.0.0"
factory-boy = "^3.1.0"
codecov = "^2.1.11"
coverage = {extras = ["toml"], version = "^5.3"}
ipython = "^7.19.0"
pre-commit = "^2.9.3"
django-coverage-plugin = "^1.8.0"
pytest-django = "^4.1.0"
python-dotenv = "^0.15.0"
django-debug-toolbar = "^3.2"
django-extensions = "^3.1.0"
click = "^7.1.2"
django-test-migrations = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tests = "scripts.tests:main"
server = "scripts.server:main"
manage = "scripts.manage:main"

[tool.coverage.run]
source = [
    "CHANGE_ME",
]
omit = [
    "tests/*",
    "factories/*"
]

[tool.coverage.report]
omit = [
    "tests/*",
    "factories/*",
    "src/CHANGE_ME/manage.py",
    "src/CHANGE_ME/wsgi.py",
]
skip_empty = true
show_missing = true
#fail_under = 70

[tool.black]
line-length = 79

[tool.pytest.ini_options]
addopts = "-s"
python_files = [
    "test_*.py",
    "tests.py",
    "*_tests.py"
]
testpaths = [
    "./tests",
]
